UPSTREAM: fscrypt: improve warning messages for unsupported encryption contexts
When fs/crypto/ encounters an inode with an invalid encryption context,
currently it prints a warning if the pair of encryption modes are
unrecognized, but it's silent if there are other problems such as
unsupported context size, format, or flags.  To help people debug such
situations, add more warning messages.

Change-Id: I7fa1288860e23ff0c30be0e8686bd824e5dc32e7
Signed-off-by: Eric Biggers <ebiggers@google.com>