UPSTREAM: ext4: fix kernel oops caused by spurious casefold flag
If an directory has the a casefold flag set without the casefold
feature set, s_encoding will not be initialized, and this will cause
the kernel to dereference a NULL pointer.  In addition to adding
checks to avoid these kernel oops, attempts to load inodes with the
casefold flag when the casefold feature is not enable will cause the
file system to be declared corrupted.

Change-Id: I9bb76d0ac3bb6709e6e9880e63300968322aace6
Signed-off-by: Theodore Ts'o <tytso@mit.edu>