UPSTREAM: fscrypt: zeroize fscrypt_info before freeing
memset the struct fscrypt_info to zero before freeing.  This isn't
really needed currently, since there's no secret key directly in the
fscrypt_info.  But there's a decent chance that someone will add such a
field in the future, e.g. in order to use an API that takes a raw key
such as siphash().  So it's good to do this as a hardening measure.

Change-Id: Ice19ef461ca621e2c6264e6503636a49a7cdcc99
Signed-off-by: Eric Biggers <ebiggers@google.com>